# GFBackend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Gator Forum Backend Server",
        "title": "Gator Forum Backend API",
        "termsOfService": "https://github.com/fongziyjun16/SE/tree/backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://167.71.166.120:10010",
    "basePath": "/gf/api",
    "paths": {
...
        "/community/create": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need community name \u0026 description only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Create a new Community",
                "parameters": [
                    {
                        "description": "Create a new community needs Creator, Name \u0026 Description.",
                        "name": "CommunityInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CommunityInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Create Community Success",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Community already exists",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/community/delete/:id": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need community id only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Create a new Community",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Community ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. No matter whether delete successfully, Return Success",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/community/get": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need page info: PageNO, pageSize only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Get Communities By Name Fuzzy Match",
                "parameters": [
                    {
                        "description": "Get Communities Info",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CommunityNameFuzzyMatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Community Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Community"
                            }
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/community/getcommunitiesbycreator": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need page info: username, PageNO, pageSize",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Get Communities By Creator",
                "parameters": [
                    {
                        "description": "Get Communities By Creator",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Community"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Community Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.NewCommunityInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/community/getcommunityidbymember": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need member name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Get Community IDs By Member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member Name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Community IDs Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/community/getmember": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need community IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Get Members By Community IDs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Community ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Members Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.CommunityMember"
                            }
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/community/getone/:id": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need community ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Get One Community By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Community ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Community Success",
                        "schema": {
                            "$ref": "#/definitions/entity.Community"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/community/getonebyname": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need community Name, page info: PageNO, pageSize",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Get Communities By Name Fuzzy Match",
                "parameters": [
                    {
                        "description": "Community Name Fuzzy Match Info",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CommunityNameFuzzyMatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Get Community Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Community"
                            }
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/community/join/:id": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need community ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community Manage"
                ],
                "summary": "Join One Community By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Community ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Join Community Success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Not Found or Existed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Community": {
            "type": "object",
            "properties": {
                "createDay": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.CommunityInfo": {
            "type": "object",
            "properties": {
                "Creator": {
                    "type": "string",
                    "example": "test1"
                },
                "Description": {
                    "type": "string",
                    "example": "this is a test community"
                },
                "Name": {
                    "type": "string",
                    "example": "community1"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.CommunityNameFuzzyMatch": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string",
                    "example": "community1"
                },
                "PageNO": {
                    "type": "integer"
                },
                "PageSize": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ResponseMsg": {
            "type": "object",
            "properties": {
                "Nickname": {
                    "type": "string",
                    "example": "James Bond"
                },
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "process successfully"
                },
                "new_community_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "cookies"
        }
    }
}